Move genomes and fitness mappings to the Population module
